Add daylight saving time edge case opts, always return zdt
Add opts to configure how to handle the edge-cases where there's either
an overlap or a gap in a daylight saving time transition. There are
three ways to handle an overlap and two to handle a gap and you'll
want different ones depending on what you're doing. Inferring this
from looking at the recex would be too hard and error prone.

Also change the return type to always be a seq of zoned date times,
instead of sometimes zdts and sometimes offset date times. Zoned date
times can have an offset like -03:00 as the zone, and and always
returning the same type seems less confusing.